+++ djangocms_page_sitemap/__init__.py
@@ -1,2 +1,5 @@
+# -*- coding: utf-8 -*-
+from __future__ import absolute_import, print_function, unicode_literals
+
 __version__ = '0.3.0.dev1'  # pragma: no cover
 __author__ = 'Iacopo Spalletti <i.spalletti@nephila.it>'  # pragma: no cover
