+++ djangocms_page_sitemap/admin.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
+from __future__ import absolute_import, print_function, unicode_literals
+
 from cms.extensions import PageExtensionAdmin
 from django.contrib import admin
