+++ djangocms_page_sitemap/settings.py
@@ -1,16 +1,18 @@
 # -*- coding: utf-8 -*-
+from __future__ import absolute_import, print_function, unicode_literals
+
 from cms.utils import get_cms_setting
 from django.conf import settings
 from django.utils.translation import ugettext_lazy as _
 
 PAGE_SITEMAP_CHANGEFREQ_DEFAULT = {
-    'always': _(u'always'),
-    'hourly': _(u'hourly'),
-    'daily': _(u'daily'),
-    'weekly': _(u'weekly'),
-    'monthly': _(u'monthly'),
-    'yearly': _(u'yearly'),
-    'never': _(u'never'),
+    'always': _('always'),
+    'hourly': _('hourly'),
+    'daily': _('daily'),
+    'weekly': _('weekly'),
+    'monthly': _('monthly'),
+    'yearly': _('yearly'),
+    'never': _('never'),
 }
 PAGE_SITEMAP_CHANGEFREQ = getattr(settings, 'PAGE_SITEMAP_CHANGEFREQ', PAGE_SITEMAP_CHANGEFREQ_DEFAULT)
 PAGE_SITEMAP_CACHE = get_cms_setting('CACHE_DURATIONS')['menus']
