+++ djangocms_page_sitemap/sitemap.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
+from __future__ import absolute_import, print_function, unicode_literals
+
 from cms.sitemaps import CMSSitemap
 from django.core.cache import cache
