+++ tests/test_toolbar.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
+from __future__ import absolute_import, print_function, unicode_literals
+
 from cms.toolbar.items import Menu, ModalItem
 from django.contrib.auth.models import Permission, User
 from django.core.urlresolvers import reverse
@@ -8,7 +10,7 @@ from django.utils.translation import ugettext_lazy as _
 from djangocms_page_sitemap.cms_toolbar import PAGE_SITEMAP_MENU_TITLE
 from djangocms_page_sitemap.models import PageSitemapProperties
 
-from . import BaseTest
+from .base import BaseTest
 
 
 class ToolbarTest(BaseTest):
@@ -91,4 +93,4 @@ class ToolbarTest(BaseTest):
         page_menu = toolbar.menus['page']
         meta_menu = page_menu.find_items(ModalItem, name="%s ..." % force_text(PAGE_SITEMAP_MENU_TITLE))[0].item
         self.assertTrue(meta_menu.url.startswith(reverse('admin:djangocms_page_sitemap_pagesitemapproperties_change', args=(page_ext.pk,))))
-        self.assertEqual(force_text(page_ext), force_text(_(u'Sitemap values for Page %s') % page1.pk))
+        self.assertEqual(force_text(page_ext), force_text(_('Sitemap values for Page %s') % page1.pk))
