+++ tests/test_utils/urls.py
@@ -1,3 +1,6 @@
+# -*- coding: utf-8 -*-
+from __future__ import absolute_import, print_function, unicode_literals
+
 from cms.utils.conf import get_cms_setting
 from django.conf import settings
 from django.conf.urls import include, patterns, url
@@ -15,7 +18,9 @@ urlpatterns = patterns(
     url(r'^media/cms/(?P<path>.*)$', 'django.views.static.serve',
         {'document_root': get_cms_setting('MEDIA_ROOT'), 'show_indexes': True}),
     url(r'^jsi18n/(?P<packages>\S+?)/$', 'django.views.i18n.javascript_catalog'),
-    url(r'^sitemap\.xml$', 'django.contrib.sitemaps.views.sitemap', {'sitemaps': {'cmspages': ExtendedSitemap}}),
+    url(r'^sitemap\.xml$', 'django.contrib.sitemaps.views.sitemap',{
+        'sitemaps': {'cmspages': ExtendedSitemap}
+    }),
 )
 
 urlpatterns += staticfiles_urlpatterns()
